Mash all filter services into one sevices
